revert(production): roll the served app back to 0fac006
Production sat on 0fac006 (2026-07-18) and was promoted straight to 197419f
on 2026-08-03 — a 16-day, 21-commit range. Users reported the live app broke
immediately afterward.

`production` is branch-protected and promote.mjs is fast-forward-only (no
--force path by design), so the rollback has to move FORWARD onto a commit
that restores the previous bytes. This restores the served surface to its
0fac006 state:

  - app.js, config.js, index.html, style.css, sw.js  -> 0fac006
  - src/*.js (the 2026-07-24 module split)           -> removed
  - tools/gen-code-map.mjs                           -> 0fac006 (tracks src/)
  - rule-usage.js + code map                         -> regenerated

Trunk keeps every commit, so nothing is lost — this only moves what the live
site serves. Re-landing the reverted work (four RENTALS dispatch fixes, three
Trips fixes, the module split, the dv2 scaffolding) is a follow-up once the
actual fault is identified.

NOTE: no defect was reproduced in the promoted build. Against production's
exact bytes, ci/smoke.mjs passes, ci/logic-test.mjs passes 706/706, all 14 ES
module imports serve 200 live, APP_ENV correctly resolves to 'production' (so
dv2 stays off), the GAS backend is healthy, and a simulated old-SW -> promote
-> reload upgrade converges with no errors. This revert is a precautionary
rollback at Jac's call, not a fix for a proven cause.
