+++ config.js
@@ -636,12 +636,6 @@ export const PERF_SAMPLE_RATE = 1;          // fraction of sessions that flush a
  * disables EXECUTION, not VISIBILITY — flagged code still ships readable in
  * the public bundle. Never gate a secret or a security/auth check on this. */
 export const FEATURES = {
-  // Phase-2 wrangler-style redesign (spec 2026-07-20 list-views-inline-expand + plan
-  // 2026-07-21-list-detail-views-build-plan). ON = the redesigned steel-canon look/surfaces
-  // (`html.dv2`), landing beside the old rendering. Ships OFF so production keeps today's look;
-  // the redesign shows AUTOMATICALLY on non-production (staging/local) for review — see the
-  // `dv2` toggle in app.js. Flip this true only when Jac approves promoting the new look live.
-  designV2: false,
   // Card-search global mode — a globe toggle inside each grid-card search bar flips
   // between per-card and whole-yard search, replacing the giant #globalsearch bar.
   // Flag ON = the globe path (old bar removed); OFF = the old #globalsearch bar.
